#387 convertImageToFiles utils
