+++ src/routes/notes/NewNote.tsx
@@ -4,7 +4,7 @@ import { useLocation } from 'react-router-dom';
 import MainContainer from '@components/_common/main-container/MainContainer';
 import NewNoteHeader from '@components/note/new-note-header/NewNoteHeader';
 import { NewNoteForm } from '@models/post';
-import { CroppedImg } from '@utils/getCroppedImg';
+import { convertImagesToFiles } from '@utils/convertImageToFiles';
 import NewNoteContent from '../../components/note/new-note-content/NewNoteContent';
 
 function NewNote() {
@@ -18,29 +18,9 @@ function NewNote() {
 
   const [noteInfo, setNoteInfo] = useState<NewNoteForm>({ content, images: [] });
 
-  const urlToFile = async (url: string, filename: string, mimeType: string): Promise<File> => {
-    const response = await fetch(url);
-    const blob = await response.blob();
-    return new File([blob], filename, { type: mimeType });
-  };
-
   useEffect(() => {
     if (location.state) {
-      const convertImagesToFiles = async () => {
-        const filePromises = images.map(async (image: string, index: number) => {
-          const file = await urlToFile(image, `image${index}.jpg`, 'image/jpeg');
-          const url = URL.createObjectURL(file);
-          return { file, url } as CroppedImg;
-        });
-
-        const fileObjects = await Promise.all(filePromises);
-        setNoteInfo((prevNoteInfo) => ({
-          ...prevNoteInfo,
-          images: fileObjects,
-        }));
-      };
-
-      convertImagesToFiles();
+      convertImagesToFiles(images, setNoteInfo);
     }
   }, [images, location.state]);
